Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
bcrypt==4.0.1
pydantic
python-multipart
scikit-learn
pandas
openai
python-dotenv